Fix some implementation defined behavior in dng_image.cpp
Some operations in dng_image.cpp of the form a + b and a - b, where a
is of type int and b is of type uint32, would be computed as unsigned
int operations due to "usual arithmetic conversions". This unsigned
int result would then be stored as an int which is implementation
defined.